refactor: cleaned up unused code, functions, and npm packages
