+++ components/Hero/Hero.tsx
@@ -1,4 +1,4 @@
-import SocialMetrics from '../SocialMetrics/SocialMetrics'
+import SocialMetrics from '../Social/SocialMetrics'
 import Intro from './Intro'
 
 export default async function Hero() {
